requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
python-dotenv
groq
pandas
numpy
pydantic
requests
python-multipart
matplotlib
jinja2
aiohttp
yfinance
